[FIX/#75] GiveAndTakeView and SaveGiftView 폴더링
